pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synmatai-test-platform"
version = "0.1.0"
description = "Automated synmatai Testing Platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.101.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "aiohttp>=3.13.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]

[tool.uv]
python-preference = "only-managed"
